Experimental branch. Added a testing directory. Added unit test for SqlPhotoStorage. Added/Updated documentation for most classes. Created skeletons for the photo gallery functionality.
The commit adds a lot of (experimental) functionality. Provides a way to
create sample photo entry objects via the ApplicationUtil class. Two sample
images are provided in the drawable folder that can be used for the photo
entries. Basically this is a prototype for how the MVC framework is to
work.