Renamed members inside SCController and PhotoGalleryController to make it more consistent. Removed the SCCommand object. Cleaned up documentation. Added dispose method to SCController.
Renamed members to there counterparts on the website (in credits on
wiki page) where I adapted the idea of a message-based MVC framework.
Removed the SCCommand class because it probably just made the messaging
mechanism slightly more confusing. Added a dispose method to SCController
because I guess it's good practice to stop any HanderThreads before quitting.